Refactor structure of code
