+++ src/main/scala/org/hanlonjohn23/models/Coordinates.scala
@@ -1,4 +1,4 @@
-package org.hanlonjohn23
+package org.hanlonjohn23.models
 
 import io.circe.Decoder.Result
 import io.circe.{Decoder, HCursor}
